+++ tests/test_tap2sna.py
@@ -354,19 +354,41 @@ class Tap2SnaTest(SkoolKitTestCase):
             create_tap_data_block([4, 5, 6]),
         ]
         tapfile = self._write_tap(blocks)
-        output, error = self.run_tap2sna(f'--tape-analysis -c first-edge=0 {tapfile}', catch_exit=0)
+        output, error = self.run_tap2sna(f'--tape-analysis -c first-edge=1 {tapfile}', catch_exit=0)
         self.assertEqual(error, '')
         exp_output = """
             T-states    EAR  Description
-                     0    0  Tone (8063 x 2168 T-states)
-              17480584    1  Pulse (667 T-states)
-              17481251    0  Pulse (735 T-states)
-              17481986    1  Data (19 bytes; 855/1710 T-states)
-              17765846    1  Pause (3500000 T-states)
-              21265846    1  Tone (3223 x 2168 T-states)
-              28253310    0  Pulse (667 T-states)
-              28253977    1  Pulse (735 T-states)
-              28254712    0  Data (5 bytes; 855/1710 T-states)
+                     1    0  Tone (8063 x 2168 T-states)
+              17480585    1  Pulse (667 T-states)
+              17481252    0  Pulse (735 T-states)
+              17481987    1  Data (19 bytes; 855/1710 T-states)
+              17765847    1  Pause (3500000 T-states)
+              21265847    1  Tone (3223 x 2168 T-states)
+              28253311    0  Pulse (667 T-states)
+              28253978    1  Pulse (735 T-states)
+              28254713    0  Data (5 bytes; 855/1710 T-states)
+        """
+        self.assertEqual(dedent(exp_output).lstrip(), output)
+
+    def test_option_tape_analysis_with_polarity(self):
+        blocks = [
+            create_tap_header_block(start=0),
+            create_tap_data_block([4, 5, 6]),
+        ]
+        tapfile = self._write_tap(blocks)
+        output, error = self.run_tap2sna(f'--tape-analysis -c polarity=1 {tapfile}', catch_exit=0)
+        self.assertEqual(error, '')
+        exp_output = """
+            T-states    EAR  Description
+                     0    1  Tone (8063 x 2168 T-states)
+              17480584    0  Pulse (667 T-states)
+              17481251    1  Pulse (735 T-states)
+              17481986    0  Data (19 bytes; 855/1710 T-states)
+              17765846    0  Pause (3500000 T-states)
+              21265846    0  Tone (3223 x 2168 T-states)
+              28253310    1  Pulse (667 T-states)
+              28253977    0  Pulse (735 T-states)
+              28254712    1  Data (5 bytes; 855/1710 T-states)
         """
         self.assertEqual(dedent(exp_output).lstrip(), output)