+++ sphinx/source/changelog.rst
@@ -12,6 +12,9 @@ Changelog
 * Added the ``load`` simulated LOAD configuration parameter to
   :ref:`tap2sna.py <tap2sna-sim-load>` (to specify an alternative command line
   to use to load the tape)
+* Added the ``polarity`` simulated LOAD configuration parameter to
+  :ref:`tap2sna.py <tap2sna-sim-load>` (to specify the EAR bit reading produced
+  by the first pulse on the tape)
 * Added support to :ref:`trace.py` for executing machine code in 128K snapshots
 * Added support to :ref:`control directive loops <ctlLoops>` for avoiding
   repetition of an ``N`` directive at the start of a loop